fix(daemon): cast rlim.Cur to uint64 for freebsd build
SageOx-Session: https://sageox.ai/repo/repo_019c5812-01e9-7b7d-b5b1-321c471c9777/sessions/2026-05-04T23-34-ryan-Ox58yT/view

+++ internal/daemon/fd_limit_unix.go
@@ -12,10 +12,14 @@ import "syscall"
 // Unix-only because syscall.Getrlimit / syscall.Rlimit / syscall.RLIMIT_NOFILE
 // are guarded by `//go:build unix` upstream and don't exist for GOOS=windows.
 // See fd_limit_windows.go for the Windows stub.
+//
+// rlim.Cur is uint64 on linux and darwin but int64 on freebsd; cast to keep
+// the cross-GOOS return type stable. Negative values can't occur for
+// RLIMIT_NOFILE in practice.
 func CurrentProcessFDLimit() uint64 {
 	var rlim syscall.Rlimit
 	if err := syscall.Getrlimit(syscall.RLIMIT_NOFILE, &rlim); err != nil {
 		return 0
 	}
-	return rlim.Cur
+	return uint64(rlim.Cur)
 }
